Check and add repository annotation
