+++ src/main/java/de/tum/cit/aet/artemis/atlas/repository/PrerequisiteRepository.java
@@ -1,11 +1,15 @@
 package de.tum.cit.aet.artemis.atlas.repository;
 
+import static de.tum.cit.aet.artemis.core.config.Constants.PROFILE_CORE;
+
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 
+import org.springframework.context.annotation.Profile;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
+import org.springframework.stereotype.Repository;
 
 import de.tum.cit.aet.artemis.atlas.domain.competency.Prerequisite;
 import de.tum.cit.aet.artemis.core.domain.Course;
@@ -14,9 +18,9 @@ import de.tum.cit.aet.artemis.core.repository.base.ArtemisJpaRepository;
 /**
  * Spring Data JPA repository for the {@link Prerequisite} entity.
  */
+@Profile(PROFILE_CORE)
+@Repository
 public interface PrerequisiteRepository extends ArtemisJpaRepository<Prerequisite, Long> {
-
-    List<Prerequisite> findAllByCourseIdOrderById(long courseId);
 
     @Query("""
             SELECT p
