+++ src/main/java/de/tum/cit/aet/artemis/core/repository/AuthorityRepository.java
@@ -1,6 +1,11 @@
 package de.tum.cit.aet.artemis.core.repository;
 
+import static de.tum.cit.aet.artemis.core.config.Constants.PROFILE_CORE;
+
 import java.util.List;
+
+import org.springframework.context.annotation.Profile;
+import org.springframework.stereotype.Repository;
 
 import de.tum.cit.aet.artemis.core.domain.Authority;
 import de.tum.cit.aet.artemis.core.repository.base.ArtemisJpaRepository;
@@ -8,6 +13,8 @@ import de.tum.cit.aet.artemis.core.repository.base.ArtemisJpaRepository;
 /**
  * Spring Data JPA repository for the Authority entity.
  */
+@Profile(PROFILE_CORE)
+@Repository
 public interface AuthorityRepository extends ArtemisJpaRepository<Authority, String> {
 
     /**
