+++ src/main/java/de/tum/cit/aet/artemis/core/repository/MigrationChangeRepository.java
@@ -1,7 +1,14 @@
 package de.tum.cit.aet.artemis.core.repository;
 
+import static de.tum.cit.aet.artemis.core.config.Constants.PROFILE_CORE;
+
+import org.springframework.context.annotation.Profile;
+import org.springframework.stereotype.Repository;
+
 import de.tum.cit.aet.artemis.core.domain.MigrationChangelog;
 import de.tum.cit.aet.artemis.core.repository.base.ArtemisJpaRepository;
 
+@Profile(PROFILE_CORE)
+@Repository
 public interface MigrationChangeRepository extends ArtemisJpaRepository<MigrationChangelog, String> {
 }
