+++ src/main/java/de/tum/cit/aet/artemis/iris/repository/IrisTextExerciseChatSessionRepository.java
@@ -1,5 +1,6 @@
 package de.tum.cit.aet.artemis.iris.repository;
 
+import static de.tum.cit.aet.artemis.core.config.Constants.PROFILE_IRIS;
 import static org.springframework.data.jpa.repository.EntityGraph.EntityGraphType.LOAD;
 
 import java.util.Collections;
@@ -7,10 +8,12 @@ import java.util.List;
 
 import jakarta.validation.constraints.NotNull;
 
+import org.springframework.context.annotation.Profile;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
+import org.springframework.stereotype.Repository;
 
 import de.tum.cit.aet.artemis.core.domain.DomainObject;
 import de.tum.cit.aet.artemis.core.exception.EntityNotFoundException;
@@ -21,6 +24,8 @@ import de.tum.cit.aet.artemis.iris.domain.session.IrisTextExerciseChatSession;
  * Repository interface for managing {@link IrisTextExerciseChatSession} entities.
  * Provides custom queries for finding text exercise chat sessions based on different criteria.
  */
+@Profile(PROFILE_IRIS)
+@Repository
 public interface IrisTextExerciseChatSessionRepository extends ArtemisJpaRepository<IrisTextExerciseChatSession, Long> {
 
     /**
