+++ src/main/java/de/tum/cit/aet/artemis/programming/repository/BuildPlanRepository.java
@@ -1,17 +1,22 @@
 package de.tum.cit.aet.artemis.programming.repository;
 
+import static de.tum.cit.aet.artemis.core.config.Constants.PROFILE_CORE;
 import static org.springframework.data.jpa.repository.EntityGraph.EntityGraphType.LOAD;
 
 import java.util.Optional;
 
+import org.springframework.context.annotation.Profile;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
+import org.springframework.stereotype.Repository;
 
 import de.tum.cit.aet.artemis.core.repository.base.ArtemisJpaRepository;
 import de.tum.cit.aet.artemis.programming.domain.ProgrammingExercise;
 import de.tum.cit.aet.artemis.programming.domain.build.BuildPlan;
 
+@Profile(PROFILE_CORE)
+@Repository
 public interface BuildPlanRepository extends ArtemisJpaRepository<BuildPlan, Long> {
 
     @Query("""
